feat: add modal window of product registration
