--- src/cart.html
@@ -1,59 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Online Store</title>
-  <link rel="stylesheet" href="main.css">
-  <link rel="stylesheet" href="template.css">
-</head>
-<body class="body">
-  <header class="header">
-    <a href="./index.html" class="header__store-logo">
-      <div class="logo__shopping-bag"></div>
-      <div class="logo__text">
-        <p class="logo__text-part">Online</p>
-        <p class="logo__text-part">Store</p>
-      </div>
-    </a>
-    <div class="header__total">
-      <p class="total__text">Cart total:</p>
-      <p class="total__amount">0.00 &euro;</p>
-    </div>
-    <div class="header__cart">
-      <div class="cart__pic"></div>
-      <div class="cart__quantity">0</div>
-    </div>
-  </header>
-  
-  <main class="main">
-    <div class="btn__wrap">
-      <button class="bay__btn">Cart is Empty</button>
-    </div>
-  </main>
-
-  <footer class="footer">
-    <div class="footer__wrap">
-      <a class="gh__person" href="https://github.com/stealthybunny" target="_blank">
-        <div class="gh__logo"></div>
-        <span class="gh__id">stealthybunny</span>  
-      </a>
-      <a class="gh__person" href="https://github.com/sssothink" target="_blank">
-        <div class="gh__logo"></div>
-        <span class="gh__id">sssothink</span>  
-      </a>
-      <div class="birth">
-        2022
-      </div>
-      <a class="rss__wrap" href="https://rs.school/js/" target="_blank">
-        <div class="rss"></div>
-      </a>
-      
-    </div>
-  </footer>
-
-
-  <script src="./modal.js"></script>
-</body>
-</html>
